[v9_10] misc x64 VS 2015 CTP fixes [#39308]
(cherry picked from commit ab973ec40cc7a79bbf930085b1293824caf73e7f)

+++ lib/isccc/cc.c
@@ -189,7 +189,7 @@ table_towire(isccc_sexpr_t *alist, isccc_region_t *target)
 	isccc_sexpr_t *kv, *elt, *k, *v;
 	char *ks;
 	isc_result_t result;
-	size_t len;
+	unsigned int len;
 
 	for (elt = isccc_alist_first(alist);
 	     elt != NULL;
@@ -198,7 +198,7 @@ table_towire(isccc_sexpr_t *alist, isccc_region_t *target)
 		k = ISCCC_SEXPR_CAR(kv);
 		ks = isccc_sexpr_tostring(k);
 		v = ISCCC_SEXPR_CDR(kv);
-		len = strlen(ks);
+		len = (unsigned int)strlen(ks);
 		INSIST(len <= 255U);
 		/*
 		 * Emit the key name.
